Create instrumentation-core apache-httpclient-4.0 (#694)
