+++ instrumentation-core/apache-httpclient-4.0/apache-httpclient-4.0.gradle
@@ -1,0 +1,24 @@
+plugins {
+  id "com.github.johnrengelman.shadow"
+}
+
+ext {
+  noShadowPublish = true
+}
+
+apply from: "$rootDir/gradle/java.gradle"
+
+group = 'io.opentelemetry.instrumentation'
+
+shadowJar {
+  archiveClassifier = 'agent'
+  configurations = []
+  relocate 'io.opentelemetry.instrumentation.apachehttpclient.v4_0', 'io.opentelemetry.auto.instrumentation.apachehttpclient.v4_0.shaded'
+}
+
+dependencies {
+  compileOnly project(':auto-bootstrap')
+  compileOnly deps.opentelemetryApi
+    
+  compileOnly group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.0'
+}
