+++ instrumentation-core/spring/spring-webflux-5.0/spring-webflux-5.0.gradle
@@ -9,7 +9,7 @@ ext {
 
 apply from: "$rootDir/gradle/java.gradle"
 
-group = 'io.opentelemetry.instrumentation.spring.webflux'
+group = 'io.opentelemetry.instrumentation'
 
 dependencies {
   compileOnly project(':auto-bootstrap')
