+++ instrumentation/apache-httpclient/apache-httpclient-4.0/apache-httpclient-4.0.gradle
@@ -30,6 +30,7 @@ testSets {
 }
 
 dependencies {
+  implementation project(path: ':instrumentation-core:apache-httpclient-4.0', configuration: 'shadow')
   compileOnly group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.0'
 
   testImplementation group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.0'
