+++ settings.gradle
@@ -150,6 +150,7 @@ include ':instrumentation:twilio-6.6'
 include ':instrumentation:vertx-3.0'
 include ':instrumentation:vertx-reactive-3.5'
 
+include ':instrumentation-core:apache-httpclient-4.0'
 include ':instrumentation-core:aws-sdk:aws-sdk-2.2'
 include ':instrumentation-core:reactor-3.1'
 include ':instrumentation-core:servlet'
